+++ src/main/scala/millfork/env/Environment.scala
@@ -433,6 +433,9 @@ class Environment(val parent: Option[Environment], val prefix: String, val cpuFa
     addThing(ConstantThing("nullptr.raw", nullptrConstant, p), None)
     addThing(ConstantThing("nullptr.raw.hi", nullptrConstant.hiByte.quickSimplify, b), None)
     addThing(ConstantThing("nullptr.raw.lo", nullptrConstant.loByte.quickSimplify, b), None)
+    val nullcharValue = options.features.getOrElse("NULLCHAR", 0L)
+    val nullcharConstant = NumericConstant(nullcharValue, 1)
+    addThing(ConstantThing("nullchar", nullcharConstant, b), None)
     val __zeropage_usage = UnexpandedConstant("__zeropage_usage", 1)
     addThing(ConstantThing("__zeropage_usage", __zeropage_usage, b), None)
     def addUnexpandedWordConstant(name: String): Unit = {